refactor : 크루 기본 이미지 반영
