+++ orury-domain/src/main/java/org/orury/domain/image/infrastructure/ImageReaderImpl.java
@@ -1,4 +1,4 @@
-package org.orury.domain.post.infrastructure;
+package org.orury.domain.image.infrastructure;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -6,7 +6,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.orury.common.error.code.FileExceptionCode;
 import org.orury.common.error.exception.InfraImplException;
 import org.orury.common.util.S3Folder;
-import org.orury.domain.global.image.ImageReader;
+import org.orury.domain.image.domain.ImageReader;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
@@ -16,8 +16,11 @@ import java.util.List;
 @Component
 @RequiredArgsConstructor
 public class ImageReaderImpl implements ImageReader {
-    @Value("${cloud.aws.s3.default-image}")
-    private String DEFAULT_IMAGE;
+    @Value("${cloud.aws.s3.default-image.user}")
+    private String USER_DEFAULT_IMAGE;
+
+    @Value("${cloud.aws.s3.default-image.crew}")
+    private String CREW_DEFAULT_IMAGE;
 
     @Value("${cloud.aws.s3.url}")
     private String URL;
@@ -27,8 +30,8 @@ public class ImageReaderImpl implements ImageReader {
      */
     @Override
     public String getImageLink(S3Folder domain, String profile) {
-        if (S3Folder.USER == domain && StringUtils.isBlank(profile)) profile = DEFAULT_IMAGE;
-        if (S3Folder.CREW == domain && StringUtils.isBlank(profile)) profile = DEFAULT_IMAGE;
+        if (S3Folder.USER == domain && StringUtils.isBlank(profile)) profile = USER_DEFAULT_IMAGE;
+        if (S3Folder.CREW == domain && StringUtils.isBlank(profile)) profile = CREW_DEFAULT_IMAGE;
         return getUrls(domain, List.of(profile)).get(0);
     }
